backend/requirements.txt
flask
requests
brainflow
matplotlib
numpy
openai
python-dotenv